# openapi.yaml
openapi: 3.0.3
info:
  title: HR Enterprise Integration (EAI) API
  version: 2.0.0
  description: |
    Dokumentasi lengkap API untuk sistem integrasi HR Enterprise yang terdiri dari
    **HRIS Service**, **Attendance Service**, **Payroll Service**, dan **Integration Service**.

    ## Enterprise Integration Patterns yang Diimplementasi
    - **Message Channel**: Queue RabbitMQ (`integration_queue`, `attendance_queue`, `payroll_queue`, `integration_dlq`)
    - **Message Translator**: Transformasi XML (Attendance) → Canonical JSON di `integration-service/translator.js`
    - **Content-Based Router**: Routing berdasarkan `eventType` di `integration-service/router.js`
    - **Message Endpoint / Adapter**: Koneksi setiap service ke RabbitMQ channel
    - **Dead Letter Channel**: DLQ `integration_dlq` via `integration_dlx` exchange untuk pesan gagal
    - **Canonical Data Model**: Format JSON standar internal (lihat schema `CanonicalMessage`)

    ## Format Data per Service
    | Service | Format Outbound ke RabbitMQ |
    |---|---|
    | HRIS Service | **JSON** (`EmployeeCreated`) |
    | Attendance Service | **XML** (`AttendanceSubmitted`) |
    | Integration Service | **Canonical JSON** (output setelah translasi) |
    | Payroll Service | Konsumen Canonical JSON |
...
paths:
...
  /employees:
    servers:
      - url: http://localhost:3001
        description: HRIS Service
    get:
      tags: [HRIS]
      operationId: hris-getEmployees
      summary: Daftar semua karyawan
      description: Mengembalikan seluruh data karyawan yang tersimpan di database HRIS.
      responses:
        '200':
          description: Daftar karyawan berhasil diambil.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
              example:
                - id: "EMP-171758"
                  name: "Budi Santoso"
                  email: "budi@enterprise.com"
                  role: "Software Engineer"
                  basicSalary: 8000000
                  published: true
                  createdAt: "2026-06-05T10:00:00.000Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [HRIS]
      operationId: hris-createEmployee
      summary: Tambah karyawan baru
      description: |
        Membuat record karyawan baru di database lokal HRIS dengan status `published: false`.
        Karyawan harus di-publish terlebih dahulu agar tersinkronisasi ke service lain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
            example:
              name: "Budi Santoso"
              email: "budi@enterprise.com"
              role: "Software Engineer"
              basicSalary: 8000000
      responses:
        '201':
          description: Karyawan berhasil ditambahkan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
              example:
                id: "EMP-171758"
                name: "Budi Santoso"
                email: "budi@enterprise.com"
                role: "Software Engineer"
                basicSalary: 8000000
                published: false
                createdAt: "2026-06-05T10:00:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /employees/{id}/publish:
    servers:
      - url: http://localhost:3001
        description: HRIS Service
    post:
      tags: [HRIS]
      operationId: hris-publishEmployee
      summary: Publish event EmployeeCreated ke RabbitMQ
      description: |
        Mengirim pesan `EmployeeCreated` dalam format **JSON** ke queue RabbitMQ `integration_queue`.
        Integration Service kemudian akan mem-forward pesan ini ke `attendance_queue` dan `payroll_queue`
        via Content-Based Router (EIP).

        **Contoh payload JSON yang dikirim ke RabbitMQ:**
        ```json
        {
          "employeeId": "EMP-171758",
          "employeeName": "Budi Santoso",
          "salary": 8000000,
          "attendanceDays": 0,
          "eventType": "EmployeeCreated",
          "timestamp": "2026-06-05T10:00:00.000Z"
        }
        ```
      parameters:
        - name: id
          in: path
          required: true
          description: ID karyawan (format EMP-XXXXXX)
          schema:
            type: string
            example: "EMP-171758"
      responses:
        '200':
          description: Event EmployeeCreated berhasil dipublish ke RabbitMQ.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Employee event published: Budi Santoso"
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─────────────────────────────────────────────────
  # ATTENDANCE SERVICE  (http://localhost:3002)
  # ─────────────────────────────────────────────────
  /attendance:
    servers:
      - url: http://localhost:3002
        description: Attendance Service
    get:
      tags: [Attendance]
      operationId: attendance-getAll
      summary: Daftar seluruh riwayat presensi
      description: Mengembalikan semua record presensi yang tersimpan di database Attendance Service.
      responses:
        '200':
          description: Riwayat presensi berhasil diambil.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attendance'
              example:
                - id: "ATT-292837"
                  employeeId: "EMP-171758"
                  employeeName: "Budi Santoso"
                  date: "2026-06-05"
                  status: "Hadir"
                  createdAt: "2026-06-05T09:00:00.000Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Attendance]
      operationId: attendance-submit
      summary: Catat presensi karyawan
      description: |
        Menyimpan data presensi, kemudian menghasilkan pesan dalam format **XML** dan
        mengirimkannya ke RabbitMQ queue `integration_queue`.

        **Contoh XML yang dikirim ke RabbitMQ (SEBELUM transformasi):**
        ```xml
        <?xml version="1.0" encoding="UTF-8"?>
        <attendance>
          <employeeId>EMP-171758</employeeId>
          <employeeName>Budi Santoso</employeeName>
          <date>2026-06-05</date>
          <status>Hadir</status>
          <eventType>AttendanceSubmitted</eventType>
          <timestamp>2026-06-05T09:00:00.000Z</timestamp>
        </attendance>
        ```

        Integration Service akan mentransformasi XML ini menjadi **Canonical JSON** (EIP: Message Translator)
        dan mem-forward-nya ke `payroll_queue`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceInput'
            example:
              employeeId: "EMP-171758"
              date: "2026-06-05"
              status: "Hadir"
      responses:
        '201':
          description: Presensi berhasil dicatat dan event XML dipublish ke RabbitMQ.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Attendance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Karyawan tidak ditemukan di database lokal Attendance. Sync dari HRIS terlebih dahulu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Employee not found locally. Please sync from HRIS."
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /attendance/employees:
    servers:
      - url: http://localhost:3002
        description: Attendance Service
    get:
      tags: [Attendance]
      operationId: attendance-getEmployees
      summary: Daftar karyawan tersinkronisasi (Attendance)
      description: |
        **Catatan:** Endpoint aslinya di server adalah `/employees`.
        Karyawan yang telah disinkronkan ke Attendance Service melalui RabbitMQ setelah event EmployeeCreated.
      responses:
        '200':
          description: Daftar karyawan lokal berhasil diambil.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "EMP-171758"
                    name:
                      type: string
                      example: "Budi Santoso"
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─────────────────────────────────────────────────
  # PAYROLL SERVICE  (http://localhost:3003)
  # ─────────────────────────────────────────────────
...
components:
  schemas:
...
    EmployeeInput:
      type: object
      required: [name, email, role, basicSalary]
      properties:
        name:
          type: string
          description: Nama lengkap karyawan
          example: "Budi Santoso"
        email:
          type: string
          format: email
          description: Alamat email unik karyawan
          example: "budi@enterprise.com"
        role:
          type: string
          description: Jabatan atau posisi karyawan
          example: "Software Engineer"
        basicSalary:
          type: number
          description: Gaji pokok per bulan (dalam Rupiah)
          example: 8000000
    AttendanceInput:
      type: object
      required: [employeeId, date, status]
      properties:
        employeeId:
          type: string
          description: ID karyawan (harus sudah ada di database Attendance)
          example: "EMP-171758"
        date:
          type: string
          format: date
          description: Tanggal presensi (YYYY-MM-DD)
          example: "2026-06-05"
        status:
          type: string
          enum: [Hadir, Sakit, Izin, Alpa]
          description: Status kehadiran karyawan. Hanya 'Hadir' yang menambah attendanceDays.
          example: "Hadir"

    # ── Response Schemas ─────────────────────────────
    Employee:
      type: object
      properties:
        id:
          type: string
          description: ID unik karyawan (auto-generated, format EMP-XXXXXX)
          example: "EMP-171758"
        name:
          type: string
          example: "Budi Santoso"
        email:
          type: string
          format: email
          example: "budi@enterprise.com"
        role:
          type: string
          example: "Software Engineer"
        basicSalary:
          type: number
          example: 8000000
        published:
          type: boolean
          description: Status apakah event EmployeeCreated sudah dikirim ke RabbitMQ
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-06-05T10:00:00.000Z"
    Attendance:
      type: object
      properties:
        id:
          type: string
          description: ID unik presensi (auto-generated, format ATT-XXXXXX)
          example: "ATT-292837"
        employeeId:
          type: string
          example: "EMP-171758"
        employeeName:
          type: string
          example: "Budi Santoso"
        date:
          type: string
          format: date
          example: "2026-06-05"
        status:
          type: string
          enum: [Hadir, Sakit, Izin, Alpa]
          example: "Hadir"
        createdAt:
          type: string
          format: date-time
          example: "2026-06-05T09:00:00.000Z"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Pesan error dari server"
  responses:
    BadRequest:
      description: Request tidak valid (field wajib kosong atau format salah).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "All fields are required"
    NotFound:
      description: Resource yang diminta tidak ditemukan.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Employee not found"
    ServiceUnavailable:
      description: Koneksi ke RabbitMQ belum siap.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "RabbitMQ connection not ready"
    InternalServerError:
      description: Terjadi kesalahan internal pada server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"